Create srussel1_Geometry Conformer and Finder Tool
This tool allows users to check and conform all normals to face outward in the scene, as well as turning on the normal pins to show which are facing outwards. Afterwards, the user will be given a dialogue box they must type, "Tris," "Quads," or "Ngons" into, and from there the tool will search and select all faces that of which are the specified geometry.